Added Documentation to main.py

+++ main.py
@@ -31,25 +31,26 @@ def forbidden(e):
 def client_error(e):
    return make_json_response({ 'error': e.description }, 400)
 
+#Get Function for all stocks at the root level. Takes nothing, because it returns all stocks
 @app.route('/', methods = ['GET'])
 def stock_list():
-    stocks = db.getStocks()
-    return make_json_response({
+    stocks = db.getStocks() #Call to get list of all stocks
+    return make_json_response({ #Respond to user with below information
         "stocks": [{
                     stock.ticker: {"price": stock.price,
                                "open": stock.openPrice,
                                "close": stock.close,
                                "last update": stock.lastUpdate}}
-            for stock in stocks
+            for stock in stocks #Do it for all stocks in stocks
         ]
         })
 
 
-
+#GET function for an individual stock
 @app.route('/<stockTicker>', methods = ['GET'])
 def stock_data(stockTicker):
-    stockTickerHelper(stockTicker)
-    return make_json_response({
+    stockTickerHelper(stockTicker) #Make sure that stock actually exists
+    return make_json_response({ #Respond with below information
         "ticker": stockTicker,
         "price": db.getStock(stockTicker).price,
         "open": db.getStock(stockTicker).openPrice,
@@ -58,36 +59,38 @@ def stock_data(stockTicker):
         "collectionID": db.getStock(stockTicker).collectionID
         })
 
-
+#Adds new stocks to the list
 @app.route('/<stockTicker>', methods = ['PUT'])
 def stock_create(stockTicker):
-    if db.getStock(stockTicker) is not None:
+    if db.getStock(stockTicker) is not None: #Lines to check to make sure no duplicates are created
         abort(403, "Stock already exists")
     db.addStock(stockTicker,None)
     db.commit()
-    return make_json_response ({'Good': 'bucket_created'}, 201)
+    return make_json_response ({'Good': 'stock created'}, 201)
 
-
+#Function Deletes a given stock
 @app.route('/<stockTicker>', methods = ['DELETE'])
 def stock_delete(stockTicker):
+    if db.getStock(stockTicker) is None: #Check that the stock being deleted actually exists
+        abort(403, "Stock does not exist")
     db.deleteStock(db.getStock(stockTicker))
     db.commit()
-    return make_json_response({}, 204)
+    return make_json_response({'Good': 'stock deleted'}, 204)
 
 
-
+#GET for a specific collection
 @app.route('/<stockTicker>/<collectionID>', methods = ['GET'])
 def collection_data(stockTicker, collectionID):
-    collectionIdHelper(collectionID)
-    return make_json_response({
+    collectionIdHelper(collectionID) #Make Sure the Collection Exists
+    return make_json_response({ #Return the following information to the user
         "id": collectionID,
         "description": db.getCollection(collectionID).description
         })
 
-
+#Creates a new collection
 @app.route('/<stockTicker>/<collectionID>', methods = ['PUT'])
 def collection_create(stockTicker,collectionID):
-    if db.getCollection(collectionID) is not None:
+    if db.getCollection(collectionID) is not None: #Check to Make sure the stock does not already exist
         abort(403, "Collection already exists")
     description = descriptionHelper()
     db.addCollection(collectionID, description)
@@ -97,28 +100,33 @@ def collection_create(stockTicker,collectionID):
 
 @app.route('/<stockTicker>/<collectionID>', methods = ['DELETE'])
 def collection_delete(stockTicker,collectionID):
+    if db.getCollection(CollectionID) is None: #Check that the collection being deleted actually exists
+        abort(403, "Collection does not exist")
     db.deleteCollection(db.getCollection(collectionID))
     db.commit()
     return make_json_response({'Good':"collection deleted"}, 204)
 
 
-
+#Responds to user query
 def make_json_response(content, response = 200, headers = {}):
    headers['Content-Type'] = 'application/json'
    return make_response(json.dumps(content), response, headers)
 
+#Makes sure that a stck actually exists
 def stockTickerHelper(ticker):
     stock = db.getStock(ticker)
     if stock == None:
         abort(404, "Stock not found")
     return stock
 
+#Makes sure that a collection actually exists
 def collectionIdHelper(id):
     collection = db.getCollection(id)
     if collection == None:
         abort(404, "Stockssssss not found")
     return collection
 
+#Gets the description for the collection
 def descriptionHelper():
    contents = request.get_json()
    if contents is None:
